move in-memory kerchunk roundtrip test to test_integration.py
